pyproject.toml
[project]
name = "echo-bot"
version = "1.0.0"
description = "A simple CLI chatbot."
authors = [
    {name = "Ron Webb",email = "ron@ronella.xyz"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pylint = "^3.3.7"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

